requirements.txt
hikari
hikari-lightbulb
python-dotenv
google-genai
pyjwt
cryptography